SMHE-2412: Remove old object config
