+++ osgp/protocol-adapter-dlms/osgp-dlms/src/main/java/org/opensmartgridplatform/dlms/objectconfig/Medium.java
@@ -2,11 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.dlmsobjectconfig.model;
+package org.opensmartgridplatform.dlms.objectconfig;
 
 public enum Medium {
   ABSTRACT,
   ELECTRICITY,
-  GAS,
-  COMBINED
+  GAS
 }
