+++ osgp/protocol-adapter-dlms/osgp-dlms/src/main/java/org/opensmartgridplatform/dlms/objectconfig/ProfileCaptureTime.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.dlmsobjectconfig.model;
+package org.opensmartgridplatform.dlms.objectconfig;
 
 public enum ProfileCaptureTime {
   QUARTER_HOUR,
