--- osgp/protocol-adapter-dlms/osgp-protocol-adapter-dlms/src/main/java/org/opensmartgridplatform/adapter/protocol/dlms/domain/commands/dlmsobjectconfig/model/DlmsAutoAnswer.java
@@ -1,15 +1,0 @@
-// SPDX-FileCopyrightText: Copyright Contributors to the GXF project
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.dlmsobjectconfig.model;
-
-import org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.dlmsobjectconfig.DlmsObjectType;
-
-public class DlmsAutoAnswer extends DlmsObject {
-  private static final int CLASS_ID_AUTO_ANSWER = 28;
-
-  public DlmsAutoAnswer(final DlmsObjectType type, final String obisCode) {
-    super(type, CLASS_ID_AUTO_ANSWER, obisCode);
-  }
-}
